Treat started segments as running when aborting.
On startup (or periodically in case of the Cassandra storage), the
RepairManager aborts segments that do not have a leader or that belong
to paused repair runs.

However, this code would only abort segments in the RUNNING state. This
is a problem because segments might also be left in the STARTED state
when a Cassandra Reaper instance is restarted. Such segments would
block other segments from running (because when checking for busy hosts,
STARTED segments are treated like RUNNING segments), but would never
finish or be restarted, thus effectively blocking the repair process.

This patch fixes this problem by treating RUNNING and STARTED segments
in the same way when aborting segments that do not have a leader or that
belong to paused repair runs.